WIP vote.js split
