+++ lib/index.js
@@ -1,6 +1,7 @@
 require('./styles');
 require('../css/main');
 require('../css/reset');
+require('./vote');
 let NewTD = require('./new-td');
 let newTDBoxCreator = require('./box-creator');
 
@@ -73,39 +74,39 @@ function checkButtons(object) {
   }
 }
 
-function upVote(arg) {
-  let $quality = arg.siblings('.quality-rating');
-  let $downButton = arg.parent().children('.down-button');
-  let $currentId = arg.parent().parent().attr('id');
-  if ($quality.text() === 'quality: swill') {
-    $quality.text('quality: plausible');
-    arg.prop('disabled', false);
-    $downButton.prop('disabled', false);
-    updateTD($currentId, 'quality','quality: plausible');
-  } else if ($quality.text() === 'quality: plausible') {
-    $quality.text('quality: genius');
-    arg.prop('disabled', true);
-    $downButton.prop('disabled', false);
-    updateTD($currentId, 'quality','quality: genius');
-  }
-}
-
-function downVote(arg) {
-  let $quality = arg.siblings('.quality-rating');
-  let $upButton = arg.parent().children('.up-button');
-  let $currentId = arg.parent().parent().attr('id');
-  if ($quality.text() === 'quality: genius') {
-    $quality.text('quality: plausible');
-    arg.prop('disabled', false);
-    $upButton.prop('disabled', false);
-    updateTD($currentId, 'quality', "quality: plausible");
-  } else if ($quality.text() === 'quality: plausible') {
-    $quality.text('quality: swill');
-    arg.prop('disabled', true);
-    $upButton.prop('disabled', false);
-    updateTD($currentId, 'quality', "quality: swill");
-  }
-}
+// function upVote(arg) {
+//   let $quality = arg.siblings('.quality-rating');
+//   let $downButton = arg.parent().children('.down-button');
+//   let $currentId = arg.parent().parent().attr('id');
+//   if ($quality.text() === 'quality: swill') {
+//     $quality.text('quality: plausible');
+//     arg.prop('disabled', false);
+//     $downButton.prop('disabled', false);
+//     updateTD($currentId, 'quality','quality: plausible');
+//   } else if ($quality.text() === 'quality: plausible') {
+//     $quality.text('quality: genius');
+//     arg.prop('disabled', true);
+//     $downButton.prop('disabled', false);
+//     updateTD($currentId, 'quality','quality: genius');
+//   }
+// }
+//
+// function downVote(arg) {
+//   let $quality = arg.siblings('.quality-rating');
+//   let $upButton = arg.parent().children('.up-button');
+//   let $currentId = arg.parent().parent().attr('id');
+//   if ($quality.text() === 'quality: genius') {
+//     $quality.text('quality: plausible');
+//     arg.prop('disabled', false);
+//     $upButton.prop('disabled', false);
+//     updateTD($currentId, 'quality', "quality: plausible");
+//   } else if ($quality.text() === 'quality: plausible') {
+//     $quality.text('quality: swill');
+//     arg.prop('disabled', true);
+//     $upButton.prop('disabled', false);
+//     updateTD($currentId, 'quality', "quality: swill");
+//   }
+// }
 
 function mainFunction(obj){
   newTDBoxCreator(obj);
